# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "handlers.traceRequest": {
        "properties": {
          "max_depth": {
            "type": "integer"
          },
          "node": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "handlers.traceResponse": {
        "properties": {
          "chain": {
            "items": {
              "$ref": "#/components/schemas/handlers.traceStep"
            },
            "type": "array"
          },
          "entry": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "handlers.traceStep": {
        "properties": {
          "depth": {
            "type": "integer"
          },
          "node": {
            "type": "string"
          },
          "via": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CSRFToken": {
        "description": "Required on write endpoints; enforced by middleware.CSRF.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "WorkspaceAuth": {
        "description": "Workspace hash via query param or JSON body; resolved by workspaceMiddleware.",
        "in": "query",
        "name": "workspace",
        "type": "apiKey"
      },
      "WorkspaceRegisteredAuth": {
        "description": "Workspace must be already-registered (write-path gate); enforced by workspaceRegisteredMiddleware.",
        "in": "query",
        "name": "workspace",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Self-describing OpenAPI 3.0 spec for nano-brain's non-MCP REST API",
    "title": "nano-brain REST API",
    "version": "1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/graph/trace": {
      "post": {
        "description": "Breadth-first traces the outgoing call chain from an entry node, up to max_depth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/handlers.traceRequest"
              }
            }
          },
          "description": "Trace query",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.traceResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "WorkspaceAuth": []
          }
        ],
        "summary": "Trace a call chain from an entry node",
        "tags": [
          "graph"
        ]
      }
    },
...
  }
}